refactored inventorytable into separate file
